ajout des suggestions de l'IA
